+++ crates/block-producer/src/batch_builder/mod.rs
@@ -236,9 +236,10 @@ impl BatchJob {
         Span::current().set_attribute("prover.kind", self.batch_prover.kind());
 
         let proven_batch = match &self.batch_prover {
-            BatchProver::Remote(prover) => {
-                prover.prove(proposed_batch).await.map_err(BuildBatchError::RemoteProverError)
-            },
+            BatchProver::Remote(prover) => prover
+                .prove(proposed_batch)
+                .await
+                .map_err(BuildBatchError::RemoteProverClientError),
             BatchProver::Local(prover) => tokio::task::spawn_blocking({
                 let prover = prover.clone();
                 move || prover.prove(proposed_batch).map_err(BuildBatchError::ProveBatchError)